+++ timeutil/timeutil_test.go
@@ -312,7 +312,7 @@ func (s *TimeUtilSuite) TestPeriod(c *C) {
 	c.Assert(p.Minutes(), Equals, 1289175)
 	c.Assert(p.Hours(), Equals, 21486)
 	c.Assert(p.Days(), Equals, 895)
-	c.Assert(p.Weeks(), Equals, 127)
+	c.Assert(p.Weeks(), Equals, 128)
 	c.Assert(p.Years(), Equals, 2)
 
 	c.Assert(Period{}.String(), Equals, "1/01/01 00:00:00 → 1/01/01 00:00:00")
@@ -378,8 +378,8 @@ func (s *TimeUtilSuite) TestHelpers(c *C) {
 	d = time.Date(2012, 1, 1, 12, 0, 0, 0, time.Local)
 	c.Assert(Since(d, DAY), Not(Equals), 0)
 
-	d = time.Date(2023, 8, 1, 12, 30, 15, 0, time.Local)
-	c.Assert(DurationAs(time.Since(d), DAY), Equals, 585)
+	d = time.Date(2023, 8, 1, 6, 30, 15, 0, time.Local)
+	c.Assert(DurationAs(time.Duration(79872232972344474), DAY), Equals, 924)
 }
 
 // ////////////////////////////////////////////////////////////////////////////////// //